% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{center_matrix}
\alias{center_matrix}
\title{Centers a matrix}
\usage{
center_matrix(x, by = "row", method = "mean", verbose = FALSE)
}
\arguments{
\item{x}{a numeric matrix or Matrix object}

\item{by}{either "row" or "col". Default is "row"}

\item{method}{either "mean" or "median". Default is "mean"}
}
\value{
A matrix with either mean or median of row/column centered around zero.
}
\description{
Centers the mean/median of each row/column of the given matrix
around zero.
}
\examples{
# Center the mean of each row around zero
m <- matrix(runif(25, 0, 100), nrow = 5, ncol = 5) # Generate a 5x5 numeric matrix
m <- center_matrix(m, by = "row", method = "mean") # Center
all(rowMeans(m) == 0) # TRUE
all(colMeans(m) == 0) # FALSE

# Center the median of each column around zero
m <- matrix(runif(25, 0, 100), nrow = 5, ncol = 5) # Generate a 5x5 numeric matrix
m <- center_matrix(m, by = "col", method = "median") # Center
all(rowMedians(m) == 0) # FALSE
all(colMedians(m) == 0) # TRUE

}
\author{
Avishay Spitzer
}
