% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{compute_complexity}
\alias{compute_complexity}
\title{Compute cell complexity}
\usage{
compute_complexity(x, return_sorted = FALSE, cell_subset = NULL,
  verbose = FALSE)
}
\arguments{
\item{x}{a numeric matrix or Matrix object}

\item{return_sorted}{should the result be returned sorted. Default is FALSE}

\item{cell_subset}{a vector of cell IDs for which the complexity should be
calculated. Default is NULL}
}
\value{
A named vector of complexity per cell ID.
}
\description{
Computes the complexity, i.e. the number of genes with count
greater than zero for each cell (column).
}
\examples{

}
\author{
Avishay Spitzer
}
