% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PreprocConfig-class}
\alias{PreprocConfig-class}
\alias{PreprocConfig}
\alias{NucseqPreprocConfig}
\alias{SS2PreprocConfig}
\alias{DefaultPreprocConfig}
\alias{complexityCutoff,PreprocConfig-method}
\alias{complexityCutoff<-,PreprocConfig-method}
\alias{expressionCutoff,PreprocConfig-method}
\alias{expressionCutoff<-,PreprocConfig-method}
\alias{housekeepingCutoff,PreprocConfig-method}
\alias{housekeepingCutoff<-,PreprocConfig-method}
\alias{logBase,PreprocConfig-method}
\alias{logBase<-,PreprocConfig-method}
\alias{scalingFactor,PreprocConfig-method}
\alias{scalingFactor<-,PreprocConfig-method}
\title{PreprocConfig class}
\usage{
PreprocConfig(complexityCutoff, expressionCutoff, housekeepingCutoff,
  logBase, scalingFactor)

NucseqPreprocConfig()

SS2PreprocConfig()

DefaultPreprocConfig()

\S4method{complexityCutoff}{PreprocConfig}(x)

\S4method{complexityCutoff}{PreprocConfig}(x) <- value

\S4method{expressionCutoff}{PreprocConfig}(x)

\S4method{expressionCutoff}{PreprocConfig}(x) <- value

\S4method{housekeepingCutoff}{PreprocConfig}(x)

\S4method{housekeepingCutoff}{PreprocConfig}(x) <- value

\S4method{logBase}{PreprocConfig}(x)

\S4method{logBase}{PreprocConfig}(x) <- value

\S4method{scalingFactor}{PreprocConfig}(x)

\S4method{scalingFactor}{PreprocConfig}(x) <- value
}
\description{
An S4 class for storing preprocessing configuration parameters.
}
\section{Functions}{
\itemize{
\item \code{PreprocConfig}: Constructs a new \code{PreprocConfig} object.

\item \code{NucseqPreprocConfig}: constructs default preprocessing configuration for data
generated from frozen samples using SmartSeq2 protocol (single-nuclei sequencing).

\item \code{SS2PreprocConfig}: constructs default preprocessing configuration for data
generated from fresh samples using SmartSeq2 protocol.

\item \code{DefaultPreprocConfig}: constructs default preprocessing configuration
(currently trhe selected default configuration is Nucseq).
}}

\section{Slots}{

\describe{
\item{\code{complexityCutoff}}{A numeric vector of length 2 representing the lower and
upper bounds of complexity (i.e. the number of detected genes per cell)}

\item{\code{expressionCutoff}}{A numeric representing the minimal log2 mean expression
per gene below which a gene is considered lowly expressed}

\item{\code{housekeepingCutoff}}{A numeric representing the log2 mean expression of
house-keeping genes (i.e. genes that are highly expressed in all cells) per
cell below which a cell is considered low quality}

\item{\code{logBase}}{A numeric representing the logarithm base for performing log
transformation on the data}

\item{\code{scalingFactor}}{A numeric representing a scaling factor by which to divide
each data point before log transformation}
}}

\section{Methods}{

\describe{
  \item{\code{complexityCutoff}}{Getter/setter for the complexity cutoff}
  \item{\code{expressionCutoff}}{Getter/setter for the expression cutoff}
  \item{\code{housekeepingCutoff}}{Getter/setter for the housekeeping cutoff}
  \item{\code{logBase}}{Getter/setter for the log base}
  \item{\code{scalingFactor}}{Getter/setter for the scaling factor}
}
}

\examples{
pc <- PreprocConfig(complexityCutoff = c(0, 10000), expressionCutoff = 5, housekeepingCutoff = 7, logBase = 2, scalingFactor = 10)

}
\author{
Avishay Spitzer
}
