% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{complexityCutoff}
\alias{complexityCutoff}
\alias{complexityCutoff<-}
\alias{expressionCutoff}
\alias{expressionCutoff<-}
\alias{housekeepingCutoff}
\alias{housekeepingCutoff<-}
\alias{scalingFactor}
\alias{scalingFactor<-}
\alias{logBase}
\alias{logBase<-}
\alias{pseudoCount}
\alias{pseudoCount<-}
\alias{typeMatrix}
\alias{typeMatrix<-}
\title{PreprocConfig generics}
\usage{
complexityCutoff(x, ...)

complexityCutoff(x, ...) <- value

expressionCutoff(x, ...)

expressionCutoff(x, ...) <- value

housekeepingCutoff(x, ...)

housekeepingCutoff(x, ...) <- value

scalingFactor(x, ...)

scalingFactor(x, ...) <- value

logBase(x, ...)

logBase(x, ...) <- value

pseudoCount(x, ...)

pseudoCount(x, ...) <- value

typeMatrix(x, ...)

typeMatrix(x, ...) <- value
}
\description{
PreprocConfig generics
}
