% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{}
\usage{
preprocess(x, complexity_cutoff, housekeeping_cutoff, expression_cutoff,
  log_base, scaling_factor, sample_id = "", cell_ids = NULL,
  forced_genes_set = NULL, use_housekeeping_filter = FALSE,
  verbose = FALSE)
}
\arguments{
\item{use_housekeeping_filter}{}
}
\value{

}
\description{

}
\details{

}
\author{
Avishay Spitzer
}
