% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{cellIDs}
\alias{cellIDs}
\alias{cellIDs<-}
\alias{geneIDs}
\alias{geneIDs<-}
\alias{statsQC}
\alias{statsQC<-}
\title{QCResults generics}
\usage{
cellIDs(object, ...)

cellIDs(object, ...) <- value

geneIDs(object, ...)

geneIDs(object, ...) <- value

statsQC(object, ...)

statsQC(object, ...) <- value
}
\description{
QCResults generics
}
