% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{scandal_cell_complexity_distribution_plot}
\alias{scandal_cell_complexity_distribution_plot}
\title{Complexity distribution plot}
\usage{
scandal_cell_complexity_distribution_plot(
  x,
  complexity_cutoff,
  sample_id,
  project_id = ".",
  show_plot = TRUE,
  save_to_file = TRUE
)
}
\arguments{
\item{x}{an expression matrix.}

\item{complexity_cutoff}{a numeric vector of length 2 representing the lower and
upper bounds of complexity (i.e. the number of detected genes per cell).}

\item{sample_id}{an ID of the sample being processed. Used for generating the filename
and hence is mandatory.}

\item{project_id}{a character identifier of the current project. Used for saving
generated plot to the project's plots directory. Default is "." (current working
directory).}

\item{show_plot}{logical indicating if the plots should be printed to the graphics
device. Allows saving the plots without printing to the graphics device by setting
to FALSE. Default is TRUE.}

\item{save_to_file}{logical indicating wether the plots should be saved to the
plots directory. Default is TRUE.}
}
\value{
Invisibly returns the generated plot.
}
\description{
Computes the complexity (number of genes with expression > 0) of each
cell in \code{x} and plots the distribution using a scatter plot. Adds two
horizontal lines demarcating the lower and upper bounds of *complexity cutoff*
parameter.
}
\seealso{
scandal_scatter_plot, compute_complexity
}
\author{
Avishay Spitzer
}
