% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-tools.R
\name{scandal_cluster_markers_density}
\alias{scandal_cluster_markers_density}
\title{Compute cluster markers density}
\usage{
scandal_cluster_markers_density(x, clusters, mname, markers,
  return_sorted = FALSE)
}
\arguments{
\item{x}{gene expression matrix}

\item{clusters}{a character/factor vector of cluster assigment for each observation (column)
in \code{x}}

\item{markers}{a character vector of marker gene symbols corresponding to subset of genes (rows)
in \code{x}}

\item{return_sorted}{wether the result should be returned in descending order. Defaule is FALSE}
}
\value{
A \lnik{tibble} with two columns: Cluster and MCD (Markers Cluster Density).
}
\description{
This function computes the expression density of a set of marker genes
for each cluster. The expression density is defined as the mean expression of the
markers gene set accross all cells belonging to the same cluster.
}
\author{
Avishay Spitzer
}
