% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_inference.R
\name{scandal_cna_classify_scores}
\alias{scandal_cna_classify_scores}
\title{Classify CNA scores}
\usage{
scandal_cna_classify_scores(cna_scores, signal_threshold = 0.05,
  correlation_threshold = 0.05, verbose = FALSE)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
This function classifies the CNA scores according to the supplied signal and correlation thresholds.
}
\seealso{
\link{scandal_cna_compute_scores}
}
\author{
Avishay Spitzer
}
