% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_inference.R
\name{scandal_cna_compute_scores}
\alias{scandal_cna_compute_scores}
\title{Compute the cell CNA score}
\usage{
scandal_cna_compute_scores(
  object,
  cna_matrix = NULL,
  hotspot_threshold = 0.9,
  verbose = FALSE
)
}
\arguments{
\item{verbose}{}
}
\value{
A \link{tibble} with three columns:
\enumerate{
  \item CellID
  \item Signal
  \item Correlation
}
}
\description{
Computes the CNA score for each cell which is composed of the CNA
signal (the sum of squares of computed CNAs) and CNA correlation (the pearson's
correlation between the cell CNA profile and the tumor's CNA profile).
}
\author{
Avishay Spitzer
}
