% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_inference.R
\name{scandal_cnv_compute_scores}
\alias{scandal_cnv_compute_scores}
\title{Compute the cell CNV score}
\usage{
scandal_cnv_compute_scores(object, cnv_matrix = NULL,
  gene_positions_table, hotspot_threshold = 0.9, verbose = FALSE)
}
\description{
Computes the CNV score for each cell which is composed of the CNV
signal (the sum of squares of computed CNVs) and CNV correlation (the pearson's
correlation between the cell CNV profile and the tumor's CNV profile).
}
