% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{scandal_complexity_distribution_plot}
\alias{scandal_complexity_distribution_plot}
\title{Plot compexity distribution accross different nodes}
\usage{
scandal_complexity_distribution_plot(
  object,
  show_plot = TRUE,
  save_to_file = TRUE
)
}
\arguments{
\item{object}{a \code{ScandalDataSet} object.}

\item{show_plot}{logical indicating if the plots should be printed to the graphics
device. Allows saving the plots without printing to the graphics device by setting
to FALSE. Default is TRUE.}

\item{save_to_file}{logical indicating wether the plots should be saved to the
plots directory. Default is TRUE.}
}
\value{
Invisibly returns the generated plot.
}
\description{
This function enables assessing the complexity distribution (which is
the most important QC metric) accross nodes pre and post QC using a box-and-whiskers
plot and visualizes the effect of preprocessing.
}
\note{
In contrast to \link{scandal_plot_qc_metrics} this function is intended to
be called **after preprocessing**.
}
\seealso{
scandal_whiskers_plot, compute_complexity
}
\author{
Avishay Spitzer
}
