% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_inference.R
\name{scandal_compute_cnv_matrix}
\alias{scandal_compute_cnv_matrix}
\title{CNV matrix computation}
\usage{
scandal_compute_cnv_matrix(x, gene_positions_table, reference_cells,
  max_genes = 5000, expression_limits = c(-3, 3), window = 100,
  scaling_factor = 0.2, initial_centering = "col",
  base_metric = "median", verbose = FALSE)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
This function
}
\details{

}
\author{
Avishay Spitzer
}
