% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_sets.R
\docType{data}
\name{SCANDAL_G2M_MARKERS}
\alias{SCANDAL_G2M_MARKERS}
\title{List of G2M cells cycle markers}
\format{An object of class \code{character} of length 65.}
\usage{
SCANDAL_G2M_MARKERS
}
\description{
List of G2M cells cycle markers
}
\keyword{datasets}
