% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{scandal_histogram_plot}
\alias{scandal_histogram_plot}
\title{Create a histogram plot}
\usage{
scandal_histogram_plot(
  data,
  title = NULL,
  xlab = NULL,
  ylab = "Frequency",
  by = 0.2,
  title_text_size = 20,
  colour = "black",
  fill = "lightblue",
  alpha = 0.2
)
}
\arguments{
\item{data}{a numeric vector of observations.}

\item{title}{an optional string for the title of the plot. Default is NULL
(no title).}

\item{xlab}{a label for the x axis. Default is NULL (no label).}

\item{ylab}{a label for the y axis. Default is "Frequency".}

\item{by}{controls the binwidth. Default is 0.2.}

\item{title_text_size}{text size of the title. Default is 20.}

\item{colour}{color of the border of the histogram bars. Default is "black".}

\item{fill}{fill color of the histogram bars. Default is "lightblue".}

\item{alpha}{transparency of the histogram bars fill color. Default is 0.2.}
}
\value{
A \link{ggplot2} object representing the histogram plot.
}
\description{
This function creates a histogram plot i.e. uses bars to visualize
the frequency distribution of \code{data}.
}
\examples{

# Generate 10K normally distributed data points
y <- rnorm(10000, mean = 5, sd = 1)

# Generate the histogram plot with bin width of 0.2
scandal_histogram_plot(y, by = .2)

library(ggplot2)

# Generate the histogram plot with bin width of 0.2. Ad the function returns a ggplot object
# we can attach to it a vertical line marking the mean of the distribution with regular
# ggplot arithmetics
scandal_histogram_plot(y, by = .2) +
    geom_vline(xintercept = mean(y), colour = "red", linetype = "dashed", size = 1)

# Same example with bin width 0.5
scandal_histogram_plot(y, by = .5) +
    geom_vline(xintercept = mean(y), colour = "red", linetype = "dashed", size = 1)

}
\author{
Avishay Spitzer
}
