% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_sets.R
\docType{data}
\name{SCANDAL_HOUSEKEEPING_GENES_LIST}
\alias{SCANDAL_HOUSEKEEPING_GENES_LIST}
\title{List of housekeeping genes}
\format{An object of class \code{character} of length 98.}
\usage{
SCANDAL_HOUSEKEEPING_GENES_LIST
}
\description{
The genes contained in the list are ubiquitously expressed
in most cells.
}
\keyword{datasets}
