% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_sets.R
\docType{data}
\name{SCANDAL_IDH_A_G1S_MARKERS}
\alias{SCANDAL_IDH_A_G1S_MARKERS}
\title{List of G1S (cell cycle) programs markers from IDH-A dataset (Astrocytoma)}
\format{An object of class \code{character} of length 40.}
\usage{
SCANDAL_IDH_A_G1S_MARKERS
}
\description{
List of G1S (cell cycle) programs markers from IDH-A dataset (Astrocytoma)
}
\keyword{datasets}
