% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_sets.R
\docType{data}
\name{SCANDAL_IDH_O_AC_MARKERS}
\alias{SCANDAL_IDH_O_AC_MARKERS}
\title{List of AC programs markers from IDH-O dataset (Oligodendroglioma)}
\format{An object of class \code{character} of length 50.}
\usage{
SCANDAL_IDH_O_AC_MARKERS
}
\description{
List of AC programs markers from IDH-O dataset (Oligodendroglioma)
}
\keyword{datasets}
