% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_sets.R
\docType{data}
\name{SCANDAL_IDH_O_G2M_MARKERS}
\alias{SCANDAL_IDH_O_G2M_MARKERS}
\title{List of G2M (cell cycle) programs markers from IDH-O dataset (Oligodendroglioma)}
\format{An object of class \code{character} of length 40.}
\usage{
SCANDAL_IDH_O_G2M_MARKERS
}
\description{
List of G2M (cell cycle) programs markers from IDH-O dataset (Oligodendroglioma)
}
\keyword{datasets}
