% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{scandal_inspect_node}
\alias{scandal_inspect_node}
\title{Node inspection}
\usage{
scandal_inspect_node(object, node_id, verbose = FALSE)
}
\arguments{
\item{object}{a \code{ScandalDataSet} object}

\item{node_id}{the ID of the node to be inspected}

\item{verbose}{suppresses all messages from this function. Default is FALSE.}
}
\description{

}
\details{

}
\examples{

}
\author{
Avishay Spitzer
}
