% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{scandal_inspect_sample}
\alias{scandal_inspect_sample}
\title{Sample inspection}
\usage{
scandal_inspect_sample(object, sample_id, node_id = NULL,
  verbose = FALSE)
}
\arguments{
\item{object}{a \code{ScandalDataSet} object}

\item{sample_id}{the ID of the sample to be inspected}

\item{verbose}{suppresses all messages from this function. Default is FALSE.}
}
\description{

}
\details{

}
\examples{

}
\author{
Avishay Spitzer
}
