% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{scandal_inspect_samples}
\alias{scandal_inspect_samples}
\title{Samples inspection}
\usage{
scandal_inspect_samples(object, sample_ids, node_id = NULL,
  verbose = FALSE)
}
\arguments{
\item{object}{a \code{ScandalDataSet} object}

\item{sample_ids}{a character vector of IDs of the samples to be inspected.}

\item{verbose}{suppresses all messages from this function. Default is FALSE.}
}
\description{

}
\details{

}
\examples{

}
\author{
Avishay Spitzer
}
