% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scandal_load_data}
\alias{scandal_load_data}
\title{Loads an RDS file from the file system}
\usage{
scandal_load_data(
  filename,
  dirname = ".",
  project_dir = ".",
  data_dir = "data",
  verbose = FALSE
)
}
\arguments{
\item{filename}{the name of the file **without .RDS suffix**.}

\item{dirname}{a sub-directory from which the object will be loaded. Default is ".",
meaning no sub-directory will be used.}

\item{project_dir}{the root directory of the current project to which all
data files are saved. Default is NULL meaning that the files will be loaded from the
working directory.}

\item{data_dir}{a subdirectory in \code{project_dir} to which all data files
are saved. Default is "data".}

\item{verbose}{suppresses all messages from this function. Default is FALSE.}
}
\value{
Returns the data object which was loaded from the RDS file.
}
\description{
Loads an RDS file from the file system, possibly from a specified sub-directory.
}
\details{
Loads an RDS file from the file system, possibly from a specified sub-directory.
}
\author{
Avishay Spitzer
}
