% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_sets.R
\docType{data}
\name{SCANDAL_MACROPHAGE_MARKERS}
\alias{SCANDAL_MACROPHAGE_MARKERS}
\title{List of macrophage markers}
\format{An object of class \code{character} of length 7.}
\usage{
SCANDAL_MACROPHAGE_MARKERS
}
\description{
List of macrophage markers
}
\keyword{datasets}
