% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{scandal_mean_expression_frequency_plot}
\alias{scandal_mean_expression_frequency_plot}
\title{Mean gene expression distribution plot}
\usage{
scandal_mean_expression_frequency_plot(x, expression_cutoff, sample_id,
  project_id = ".", show_plot = TRUE, save_to_file = TRUE,
  histogram_bin_width = 0.2)
}
\arguments{
\item{x}{an expression matrix.}

\item{expression_cutoff}{a numeric representing the minimal log2 mean expression
per gene below which a gene is considered lowly expressed.}

\item{sample_id}{an ID of the sample being processed. Used for generating the filename
and hence is mandatory.}

\item{project_id}{a character identifier of the current project. Used for saving
generated plot to the project's plots directory. Default is "." (current working
directory).}

\item{show_plot}{logical indicating if the plots should be printed to the graphics
device. Allows saving the plots without printing to the graphics device by setting
to FALSE. Default is TRUE.}

\item{save_to_file}{logical indicating wether the plots should be saved to the
plots directory. Default is TRUE.}

\item{histogram_bin_width}{controls the width of each histogram bin. Default is 0.2.}
}
\value{
Invisibly returns the generated plot.
}
\description{
Computes the *log2(mean exression)* of each genes in \code{x} and
plots a  histogram depicting the distribution mean gene expression. Adds a
vertical line demarcating the *expression cutoff*.
}
\seealso{
scandal_histogram_plot
}
\author{
Avishay Spitzer
}
