% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{scandal_mean_housekeeping_expression_plot}
\alias{scandal_mean_housekeeping_expression_plot}
\title{Housekeeping genes expression plot}
\usage{
scandal_mean_housekeeping_expression_plot(x, housekeeping_cutoff,
  sample_id, project_id = ".", show_plot = TRUE, save_to_file = TRUE)
}
\arguments{
\item{x}{an expression matrix.}

\item{housekeeping_cutoff}{a numeric representing the log2 mean expression of
house-keeping genes (i.e. genes that are highly expressed in all cells) per
cell below which a cell is considered low quality.}

\item{sample_id}{an ID of the sample being processed. Used for generating the filename
and hence is mandatory.}

\item{project_id}{a character identifier of the current project. Used for saving
generated plot to the project's plots directory. Default is "." (current working
directory).}

\item{show_plot}{logical indicating if the plots should be printed to the graphics
device. Allows saving the plots without printing to the graphics device by setting
to FALSE. Default is TRUE.}

\item{save_to_file}{logical indicating wether the plots should be saved to the
plots directory. Default is TRUE.}
}
\value{
Invisibly returns the generated plot.
}
\description{
Computes the *log2(mean expression)* of housekeeping genes for
each cell in \code{x} and plots the distribution using a scatter plot. Adds a
horizontal line demarcating the *housekeeping cutoff* parameter.
}
\seealso{
scandal_scatter_plot, SCANDAL_HOUSEKEEPING_GENES_LIST
}
\author{
Avishay Spitzer
}
