% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_sets.R
\docType{data}
\name{SCANDAL_OLIGODENDROCYTE_MARKERS}
\alias{SCANDAL_OLIGODENDROCYTE_MARKERS}
\title{List of oligodendrocyte markers}
\format{An object of class \code{character} of length 8.}
\usage{
SCANDAL_OLIGODENDROCYTE_MARKERS
}
\description{
List of oligodendrocyte markers
}
\keyword{datasets}
