% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{scandal_plot_qc_metrics}
\alias{scandal_plot_qc_metrics}
\title{Plot quality control metrics}
\usage{
scandal_plot_qc_metrics(object, show_plot = TRUE, save_to_file = TRUE,
  histogram_bin_width = 0.2)
}
\arguments{
\item{object}{a \code{ScandalDataSet} object.}

\item{show_plot}{logical indicating if the plots should be printed to the graphics
device. Allows saving the plots without printing to the graphics device by setting
to FALSE. Default is TRUE.}

\item{save_to_file}{logical indicating wether the plots should be saved to the
plots directory. Default is TRUE.}

\item{histogram_bin_width}{controls the width of each histogram bin. Default is 0.2.}
}
\description{
This function generates 3 types of plots for each node in \code{object}:
\enumerate{
  \item A scatter plot depicting the distribution of complexities within each node.
  \item A scatter plot depicting the mean expression of housekeeping gene for each cell.
  \item A hostogram plot depicting the mean expression distribution of each gene.
}
The function is inteneded to be called **prior** to preproccesing to assist in choosing
the right QC thresholds for each metric and enable plotting and saving in bulk for all
the nodes contained in \code{object}.
}
\seealso{
plot_cell_complexity_distribution, plot_mean_housekeeping_expression,
plot_mean_expression_frequency
}
\author{
Avishay Spitzer
}
