% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{scandal_preprocess}
\alias{scandal_preprocess}
\title{}
\usage{
scandal_preprocess(object, preproc_config_list, forced_genes_set = NULL,
  use_housekeeping_filter = FALSE, verbose = FALSE)
}
\arguments{
\item{use_housekeeping_filter}{}
}
\value{

}
\description{

}
\details{

}
\author{
Avishay Spitzer
}
