% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{scandal_programs_of_intra_sample_heterogeneity}
\alias{scandal_programs_of_intra_sample_heterogeneity}
\title{Programs of intra-sample heterogeneity}
\usage{
scandal_programs_of_intra_sample_heterogeneity(object, samples = NULL,
  clustering_data = NULL, algorithm = "nmf", rank = 10,
  ngenes1 = 50, ngenes2 = 30, sd_threshold = 0.8,
  filter_method = "relative", bin_control = TRUE,
  n_control_bins = 25, n_bin_genes = 100, return_all = FALSE,
  verbose = FALSE, ...)
}
\arguments{
\item{...}{further arguments passed to nmf function}
}
\value{

}
\description{
This function extracts transpriptomic programs
}
\details{
Generally, the algorithm extracts programs in a bottoms-up approach
starting from extracting programs within each individual sample. The algorithm
then detecs highly variable programs that best represent coherent program clusters
and aggregates these clusters into programs that generalize accross all samples.
\cr
The algorithm performs the following steps:
\cr
\enumerate{
  \item Call \link{prepare_samples} to prepare a \linkS4class{ScandalDataSet} for
  each individual sample. This step can be bypassed by supplying a valid \code{samples}
  argument.
  \item For each individual sample generate a predefined number of clusters
  (configurable by the \code{rank} parameter). At the moment the only supported
  clustering algorithm is NMF via \link{nmf_run}. This step can be bypassed by supplying
  a valid \code{clustering_data} argument.
  \item For each sample extract an initial list of within-sample programs by calling
  \link{nmf_extract_programs}. The number of genes in each program is configurable by
  the \code{ngenes1} parameter. The number of initial within-sample programs correpsonds
  to \code{rank}.
  \item For each sample score the cells for each of the within-sample programs by calling
  \link{score_within_samples}. See \link{scrabble::score} for more details about how to
  score cells while controling for differences in cell complexities.
  \item For each sample compute the standard deviation of scores for each within-sample
  program by calling \link{compute_programs_sd}.
  \item
}
\cr
}
\seealso{

}
\author{
Avishay Spitzer
}
