% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{scandal_quality_control_stats}
\alias{scandal_quality_control_stats}
\title{Quality control statistics}
\usage{
scandal_quality_control_stats(object)
}
\arguments{
\item{object}{a \code{ScandalDataSet} object}
}
\description{
This function returns a \code{DataFrame} containing statistics
gathered while performing the preprocessing procedure including cell and gene
drop rates.
}
