% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{scandal_reduced_dims_plot}
\alias{scandal_reduced_dims_plot}
\title{Create a reduced dimensions plot}
\usage{
scandal_reduced_dims_plot(
  object,
  rdims,
  labels = NULL,
  legend_name = NULL,
  title = DEFAULT_TITLE(object, "Reduced dimensions plot"),
  title_text_size = 20,
  xlab = "dim1",
  ylab = "dim2"
)
}
\arguments{
\item{object}{a \linkS4class{ScandalDataSet} object.}

\item{rdims}{the name of the reduced dimensions object that will be retrieved from
\code{object} (using \code{reducedDim} method).}

\item{labels}{an optional vector of character labels for each (x, y) point.
Used for color coding each point. Default is NULL.}

\item{legend_name}{the name of the leged. Default is NULL (no label).}

\item{title}{an optional string for the title of the plot. Default is NULL
(no title).}

\item{title_text_size}{text size of the title. Default is 20.}
}
\value{
A \link{ggplot2} object representing the reduced dimensions plot.
}
\description{
This function plots the reduced dimensions coordinates using a scatter plot.
}
\seealso{
\link{umap}
}
\author{
Avishay Spitzer
}
