% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scandal_save_data}
\alias{scandal_save_data}
\title{Saves a data object as an RDS file}
\usage{
scandal_save_data(
  object,
  filename,
  dirname = ".",
  project_dir = ".",
  data_dir = "data",
  verbose = FALSE
)
}
\arguments{
\item{object}{an object to be saved as a RDS file.}

\item{filename}{the name of the file **without .RDS suffix**.}

\item{dirname}{a sub-directory to which the object will be saved. Default is ".",
meaning no sub-directory will be used.}

\item{project_dir}{the root directory of the current project to which all
data files are saved. Default is NULL meaning that the files will be saved to the
working directory.}

\item{data_dir}{a subdirectory in \code{project_dir} to which all data files
are saved. Default is "data".}

\item{verbose}{suppresses all messages from this function. Default is FALSE.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Saves a data object as an RDS files, possibly into a specified sub-directory.
Creates the sub-directory if it is missing.
}
\details{
Saves a data object as an RDS files, possibly into a specified sub-directory.
Creates the sub-directory if it is missing.
}
\author{
Avishay Spitzer
}
