% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{scandal_scatter_plot}
\alias{scandal_scatter_plot}
\title{Create a scatter plot}
\usage{
scandal_scatter_plot(
  x,
  y,
  labels = NULL,
  color_legend_name = NULL,
  title = NULL,
  xlab = NULL,
  ylab = NULL,
  plot_ordered = FALSE,
  order_by_axis = "y",
  title_text_size = 20
)
}
\arguments{
\item{x}{a numeric vector representing the variable to plot in the x axis. If
set to NULL then a vector \code{1:length(y)} will be used instead.}

\item{y}{a numeric vector representing the variable to plot in the y axis. If
set to NULL then a vector \code{1:length(x)} will be used instead.}

\item{labels}{an optional vector of character labels for each (x, y) point.
Used for color coding each point. Default is NULL.}

\item{color_legend_name}{the name of the legend, valid only if the labels vector
is provided. Default is NULL (can be set to NULL as well if labels vector
is provided).}

\item{title}{an optional string for the title of the plot. Default is NULL
(no title).}

\item{xlab}{a label for the x axis. Default is NULL (no label).}

\item{ylab}{a label for the y axis. Default is NULL (no label).}

\item{plot_ordered}{a logical indicating of the provided data should be
plotted ordered. Valid only if either \code{x} or \code{y} are provided
but not both of them. Default is FALSE.}

\item{order_by_axis}{the axis by which to order the data (either "x" or "y").
Default is "y".}

\item{title_text_size}{text size of the title. Default is 20.}
}
\value{
A \link{ggplot2} object representing the scatter plot.
}
\description{
This function creates a scatter plot i.e. a two-dimensional plot
that uses dots to visualize the values of two different variables \code{x} and
\code{y}.
}
\examples{

# Example #1 - One variable of interest

# Generate randomly 1000 data points
y <- rnorm(1000)

# See the difference between the ordered and unordered scatter plots
scandal_scatter_plot(x = NULL, y = y, plot_ordered = TRUE, order_by_axis = "y", title = "Ordered plot")
scandal_scatter_plot(x = NULL, y = y, plot_ordered = FALSE, title = "Unordered plot")

# Example #2 - Two variables with linear correlation

library(MASS)
library(ggplot2)

samples <- 1000
r <- 0.8

# Generate 2 variables with a linear correlation between them (1000 data points with pearsnon's r 0.8)
data <- mvrnorm(n = samples, mu = c(0, 0), Sigma = matrix(c(1, r, r, 1), nrow = 2), empirical = TRUE)

# Plot the two variables. As scandal_scatter_plot returns a ggplot object we can add to it a
# linear regression line showing the positive correlation
scandal_scatter_plot(x = data[, 1], y = data[, 2], plot_ordered = FALSE) +
    geom_smooth(method = "glm")

}
\author{
Avishay Spitzer
}
