% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-tools.R
\name{scandal_setup_analysis}
\alias{scandal_setup_analysis}
\title{Setup analysis tools}
\usage{
scandal_setup_analysis(
  object,
  computation_opts = c(correlation = TRUE, tsne = TRUE, umap = TRUE),
  cor_method = "pearson",
  tsne_perplexity = 30,
  tsne_pca = TRUE,
  tsne_initial_dims = 50,
  tsne_pca_center = TRUE,
  tsne_pca_scale = FALSE,
  tsne_normalize = TRUE,
  umap_config = scandal_default_umap_config(),
  verbose = FALSE
)
}
\arguments{
\item{umap_config}{}
}
\description{
This function sets-up analysis-related tools such as computing a cell-correlation
matrix, t-SNE and UMAP to be used later in downstream analysis.
}
\author{
Avishay Spitzer
}
