% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{scandal_tsne_plot}
\alias{scandal_tsne_plot}
\title{Create a t-SNE plot}
\usage{
scandal_tsne_plot(object, tsne_labels = NULL, legend_name = NULL,
  title = DEFAULT_TITLE(object, "t-SNE plot"), title_text_size = 20)
}
\arguments{
\item{object}{a \code{\links4class{ScandalDataSet}} object.}

\item{tsne_labels}{an optional vector of character labels for each (x, y) point.
Used for color coding each point. Default is NULL.}

\item{legend_name}{the name of the leged. Default is NULL (no label).}

\item{title}{an optional string for the title of the plot. Default is NULL
(no title).}

\item{title_text_size}{text size of the title. Default is 20.}
}
\value{
A \link{ggplot2} object representing the histogram plot.
}
\description{
This function plots the t-SNE coordinates using a scatter plot.
}
\seealso{
\link{Rtsne}
}
\author{
Avishay Spitzer
}
