% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{scandal_umap_plot}
\alias{scandal_umap_plot}
\title{Create a UMAP plot}
\usage{
scandal_umap_plot(
  object,
  umap_labels = NULL,
  legend_name = NULL,
  title = DEFAULT_TITLE(object, "UMAP plot"),
  title_text_size = 20
)
}
\arguments{
\item{object}{a \linkS4class{ScandalDataSet} object.}

\item{umap_labels}{an optional vector of character labels for each (x, y) point.
Used for color coding each point. Default is NULL.}

\item{legend_name}{the name of the leged. Default is NULL (no label).}

\item{title}{an optional string for the title of the plot. Default is NULL
(no title).}

\item{title_text_size}{text size of the title. Default is 20.}
}
\value{
A \link{ggplot2} object representing the UMAP plot.
}
\description{
This function plots the UMAP coordinates using a scatter plot.
}
\seealso{
\link{umap}
}
\author{
Avishay Spitzer
}
