% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{scandal_whiskers_plot}
\alias{scandal_whiskers_plot}
\title{Create a box-and-whiskers plot}
\usage{
scandal_whiskers_plot(data, labels, title = NULL, xlab = NULL,
  ylab = NULL, palette = NULL, legend_name = NULL,
  title_text_size = 20)
}
\arguments{
\item{data}{a numeric vector of observations.}

\item{labels}{a vector of label per observation in \code{data} used to spilt
\code{data} into the different groups.}

\item{title}{an optional string for the title of the plot. Default is NULL
(no title).}

\item{xlab}{a label for the x axis. Default is NULL (no label).}

\item{ylab}{a label for the y axis. Default is NULL (no label).}

\item{palette}{an optional color palette used instead of \code{ggplot}'s
default palette. Default is NULL.}

\item{legend_name}{the name of the leged. Default is NULL (no label).}

\item{title_text_size}{text size of the title. Default is 20.}
}
\value{
A \link{ggplot2} object representing the histogram plot.
}
\description{
This function depicts the distribution of data within groups
using a box and whiskers diagram. Each box depicts the interquantile
distribution within a group while the whiskers depict the distribution in
the upper and lower quartiles.
}
\examples{

# Generate 1K normally distributed data points
y <- rnorm(1000, mean = 5, sd = 1)

# Randomlly assign each data point to a group
labels <- paste0("Group", sample(x = 1:5, size = 1000, replace = TRUE))

# Plot the distribution
scandal_whiskers_plot(y, labels)

}
\author{
Avishay Spitzer
}
