% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ScandalDataSet-class}
\alias{ScandalDataSet-class}
\title{ScandalDataSet class}
\description{
An S4 class for storing single-cell seqeuncing data, analysis
}
\section{Slots}{

\describe{
\item{\code{projectID}}{}
}}

\author{
Avishay Spitzer
}
