% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ScandalDataSet-class}
\alias{ScandalDataSet-class}
\alias{ScandalDataSet}
\alias{logtpm,ScandalDataSet-method}
\alias{logtpm<-,ScandalDataSet-method}
\alias{qualityControl,ScandalDataSet-method}
\alias{qualityControl<-,ScandalDataSet-method}
\alias{nodeID,ScandalDataSet-method}
\alias{projectID,ScandalDataSet-method}
\alias{nodeIDs,ScandalDataSet-method}
\alias{unprocessedData,ScandalDataSet-method}
\alias{preprocConfig,ScandalDataSet-method}
\alias{inspectNode,ScandalDataSet-method}
\title{ScandalDataSet class}
\usage{
## Constructor
ScandalDataSet(..., preprocConfig = DefaultPreprocConfig(),
  nodeID = NODE_ID(), projectID = PROJ_ID())

\S4method{logtpm}{ScandalDataSet}(object, ...)

\S4method{logtpm}{ScandalDataSet}(object, ...) <- value

\S4method{qualityControl}{ScandalDataSet}(object)

\S4method{qualityControl}{ScandalDataSet}(object) <- value

\S4method{nodeID}{ScandalDataSet}(object)

\S4method{projectID}{ScandalDataSet}(object)

\S4method{nodeIDs}{ScandalDataSet}(object)

\S4method{unprocessedData}{ScandalDataSet}(object)

\S4method{preprocConfig}{ScandalDataSet}(object)

\S4method{inspectNode}{ScandalDataSet}(object, nodeID)
}
\arguments{
\item{...}{arguments to pass to the \code{SingleCellExperiment} constructor.}

\item{preprocConfig}{a configuration object of class \code{PreprocConfig}}

\item{nodeID}{a unique identifier of the constructed \code{ScandalDataSet}
object that should represent the specific sample. If not supplied a unique ID
will be generated randomly however it is advised to set this field.}

\item{projectID}{an identifier common to all the nodes in the constructed
\code{ScandalDataSet} object.  If not supplied a unique ID
will be generated randomly however it is advised to set this field.}

\item{value}{a value to replace the currently set value (applies to all setter methods).}
}
\description{
An S4 class for storing single-cell seqeuncing data, analysis
}
\details{
The S4 class \code{ScandalDataSet}
}
\section{Slots}{

\describe{
\item{\code{unprocessedData}}{a read-only matrix that contains the unprocessed data that
allows re-accessing this data without the need to read it from file. Sparse matrix
representation as well as maintaining a single copy for the entire objects tree
decreases the memory overhead of this approach.}

\item{\code{preprocConfig}}{a configuration object of class \code{PreprocConfig}.}

\item{\code{qualityControl}}{a \code{SimpleList} object containing objects of class
\code{QCResults} representing the quality control results, i.e. the cells and
genes in each individual node that passed qaulity control and are available for
downstream analysis.}

\item{\code{nodeID}}{a unique character identifier of the constructed \code{ScandalDataSet}
object that should represent the specific sample.}

\item{\code{projectID}}{a character identifier common to all the nodes in the constructed
\code{ScandalDataSet} object.}
}}

\section{Constructor}{

Constructs a \code{ScandalDataSet} object.
}

\section{Methods}{

\describe{
  \item{\code{logtpm}}{Getter/setter for the logtpm assay}
  \item{\code{unprocessedData}}{Getter for the unprocessedData (read-only)}
  \item{\code{preprocConfig}}{Getter for the preprocConfig (read-only)}
  \item{\code{qualityControl}}{Getter/Setter for the qualityControl list}
  \item{\code{nodeID}}{Getter/setter for the nodeID}
  \item{\code{projectID}}{Getter/setter for the projectID}
  \item{\code{nodeIDs}}{Returns a character vector containing the IDs of all
  nodes (samples) in the dataset.}
  \item{\code{inspectNode}}{Returns a ScandalDataSet object representing a
  specific node (sample).}
}
}

\examples{



}
\author{
Avishay Spitzer
}
