% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{center_matrix}
\alias{center_matrix}
\title{Centers a matrix}
\usage{
center_matrix(x, by = "row", method = "mean", verbose = FALSE)
}
\arguments{
\item{x}{a numeric matrix}

\item{method}{the}
}
\value{

}
\description{
Centers each row or column of the given matrix around the mean or median
}
\details{

}
\author{
Avishay Spitzer
}
