% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{load_dataset}
\alias{load_dataset}
\title{Load dataset from file}
\usage{
load_dataset(filename, drop_cols = 2, rownames_col = 1,
  excluded_samples = NULL, verbose = FALSE)
}
\arguments{
\item{excluded_samples}{}
}
\value{

}
\description{

}
\details{

}
\author{
Avishay Spitzer
}
