% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{load_dataset}
\alias{load_dataset}
\title{Load dataset from file}
\usage{
load_dataset(
  filename,
  cell_2_node_map = DEFAULT_CELL_2_NODE_MAP,
  drop_cols = 1,
  rownames_col = 1,
  excluded_samples = NULL,
  as_Matrix = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{filename}{the name of the file containing the dataset (in tab-delimited
format).}

\item{cell_2_node_map}{a **function** that maps a vector of cell IDs to a vector of
node IDs to which the cells belong. The default function assumes that the cell ID
is a string separated by "-" and that the node ID is contained in the substring
until the first "-" character.}

\item{drop_cols}{number of columns that should be dropped from the dataset,
i.e. if drop_cols == 3 then columns 1:3 will be dropped. Default is 1.}

\item{rownames_col}{the column that contains the rownames, i.e. gene symbols
or identifiers. Default is 1.}

\item{excluded_samples}{a vector containing the names of the samples that
should be excluded from the returned dataset. Defaults to NULL.}

\item{as_Matrix}{logical indicating whether the loaded dataset should be returned
as an S4 Matrix class (supports sparse representation) or base R matrix type.
Default is TRUE.}

\item{verbose}{suppresses all messages from this function. Default is FALSE.}
}
\value{

}
\description{
Loads a dataset from file
}
\details{

}
\examples{

}
\author{
Avishay Spitzer
}
