% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PreprocConfig-class}
\alias{PreprocConfig-class}
\alias{PreprocConfig}
\alias{NucseqPreprocConfig}
\alias{SS2PreprocConfig}
\alias{DefaultPreprocConfig}
\alias{complexityCutoff,PreprocConfig-method}
\alias{complexityCutoff<-,PreprocConfig-method}
\alias{expressionCutoff,PreprocConfig-method}
\alias{expressionCutoff<-,PreprocConfig-method}
\alias{housekeepingCutoff,PreprocConfig-method}
\alias{housekeepingCutoff<-,PreprocConfig-method}
\alias{logBase,PreprocConfig-method}
\alias{logBase<-,PreprocConfig-method}
\alias{scalingFactor,PreprocConfig-method}
\alias{scalingFactor<-,PreprocConfig-method}
\alias{pseudoCount,PreprocConfig-method}
\alias{pseudoCount<-,PreprocConfig-method}
\alias{typeMatrix,PreprocConfig-method}
\alias{typeMatrix<-,PreprocConfig-method}
\alias{show,PreprocConfig-method}
\title{PreprocConfig class}
\usage{
PreprocConfig(complexityCutoff, expressionCutoff, housekeepingCutoff,
  logBase, scalingFactor, pseudoCount, typeMatrix)

NucseqPreprocConfig()

SS2PreprocConfig()

DefaultPreprocConfig()

\S4method{complexityCutoff}{PreprocConfig}(x, sid = NULL)

\S4method{complexityCutoff}{PreprocConfig}(x) <- value

\S4method{expressionCutoff}{PreprocConfig}(x, sid = NULL)

\S4method{expressionCutoff}{PreprocConfig}(x) <- value

\S4method{housekeepingCutoff}{PreprocConfig}(x, sid = NULL)

\S4method{housekeepingCutoff}{PreprocConfig}(x) <- value

\S4method{logBase}{PreprocConfig}(x)

\S4method{logBase}{PreprocConfig}(x) <- value

\S4method{scalingFactor}{PreprocConfig}(x)

\S4method{scalingFactor}{PreprocConfig}(x) <- value

\S4method{pseudoCount}{PreprocConfig}(x)

\S4method{pseudoCount}{PreprocConfig}(x) <- value

\S4method{typeMatrix}{PreprocConfig}(x)

\S4method{typeMatrix}{PreprocConfig}(x) <- value

\S4method{show}{PreprocConfig}(object)
}
\arguments{
\item{complexityCutoff}{A numeric vector of length 2 representing the lower and
upper bounds of complexity (i.e. the number of detected genes per cell).}

\item{expressionCutoff}{A numeric representing the minimal log2 mean expression
per gene below which a gene is considered lowly expressed.}

\item{housekeepingCutoff}{A numeric representing the log2 mean expression of
house-keeping genes (i.e. genes that are highly expressed in all cells) per
cell below which a cell is considered low quality.}

\item{logBase}{A numeric representing the logarithm base for performing log
transformation on the data.}

\item{scalingFactor}{A numeric representing a scaling factor by which to divide
each data point before log transformation.}

\item{pseudoCount}{A numeric representing the pseudo count added when performing
log transformation to avoid taking the log of zero.}

\item{typeMatrix}{A logical indicating if the dataset should be represented using
the S4 Matrix class (instead of base R matrix) to reduce memory overhead using
sparse matrix representation.}

\item{x}{a \code{PreprocConfig} object.}

\item{value}{a value to replace the currently set value.}
}
\description{
An S4 class for storing preprocessing configuration parameters.
}
\section{Functions}{
\itemize{
\item \code{PreprocConfig}: Constructs a new \code{PreprocConfig} object.

\item \code{NucseqPreprocConfig}: constructs default preprocessing configuration for data
generated from frozen samples using SmartSeq2 protocol (single-nuclei sequencing).

\item \code{SS2PreprocConfig}: constructs default preprocessing configuration for data
generated from fresh samples using SmartSeq2 protocol.

\item \code{DefaultPreprocConfig}: constructs default preprocessing configuration
(currently trhe selected default configuration is Nucseq).
}}

\section{Slots}{

\describe{
\item{\code{complexityCutoff}}{A numeric vector of length 2 representing the lower and
upper bounds of complexity (i.e. the number of detected genes per cell).}

\item{\code{expressionCutoff}}{A numeric representing the minimal log2 mean expression
per gene below which a gene is considered lowly expressed.}

\item{\code{housekeepingCutoff}}{A numeric representing the log2 mean expression of
house-keeping genes (i.e. genes that are highly expressed in all cells) per
cell below which a cell is considered low quality.}

\item{\code{logBase}}{A numeric representing the logarithm base for performing log
transformation on the data.}

\item{\code{scalingFactor}}{A numeric representing a scaling factor by which to divide
each data point before log transformation.}

\item{\code{pseudoCount}}{A numeric representing the pseudo count added when performing
log transformation to avoid taking the log of zero.}

\item{\code{typeMatrix}}{A logical indicating if the dataset should be represented using
the S4 Matrix class (instead of base R matrix) to reduce memory overhead using
sparse matrix representation.}
}}

\section{Methods}{

\describe{
  \item{\code{complexityCutoff}}{Getter/setter for the complexity cutoff}
  \item{\code{expressionCutoff}}{Getter/setter for the expression cutoff}
  \item{\code{housekeepingCutoff}}{Getter/setter for the housekeeping cutoff}
  \item{\code{logBase}}{Getter/setter for the log base}
  \item{\code{scalingFactor}}{Getter/setter for the scaling factor}
  \item{\code{pseudoCount}}{Getter/setter for the pseudo count}
  \item{\code{typeMatrix}}{Getter/setter for the Matrix type}
}
}

\examples{
pc <- PreprocConfig(complexityCutoff = c(0, 10000),
                    expressionCutoff = 5,
                    housekeepingCutoff = 7,
                    logBase = 2,
                    scalingFactor = 10,
                    pseudoCount = 1,
                    typeMatrix = TRUE)

logBase(pc) # Equals 2
logBase(pc) <- 10
logBase(pc) # Equals 10

}
\author{
Avishay Spitzer
}
