% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{QCResults-class}
\alias{QCResults-class}
\alias{QCResults}
\alias{preprocConfig,QCResults-method}
\alias{cellIDs,QCResults-method}
\alias{cellIDs<-,QCResults-method}
\alias{geneIDs,QCResults-method}
\alias{geneIDs<-,QCResults-method}
\alias{statsQC,QCResults-method}
\alias{statsQC<-,QCResults-method}
\title{QCResults class}
\usage{
QCResults(preprocConfig, cellIDs = c(), geneIDs = c(),
  statsQC = S4Vectors::DataFrame())

\S4method{preprocConfig}{QCResults}(object)

\S4method{cellIDs}{QCResults}(object)

\S4method{cellIDs}{QCResults}(object) <- value

\S4method{geneIDs}{QCResults}(object)

\S4method{geneIDs}{QCResults}(object) <- value

\S4method{statsQC}{QCResults}(object)

\S4method{statsQC}{QCResults}(object) <- value
}
\description{
An S4 class for storing quality control results.
}
\section{Functions}{
\itemize{
\item \code{QCResults}: Constructs a new \code{QCResults} object.
}}

\section{Slots}{

\describe{
\item{\code{preprocConfig}}{A object of class \code{PreprocConfig}.}

\item{\code{cellIDs}}{A vector.}

\item{\code{geneIDs}}{A vector.}

\item{\code{statsQC}}{A DataFrame.}
}}

\section{Methods}{

\describe{
  \item{\code{preprocConfig}}{Getter for the PreprocConfig used when creating
  the quality control result.}
  \item{\code{cellIDs}}{Getter for the cell IDs that passed QC.}
  \item{\code{geneIDs}}{Getter for the gene IDs that passed QC}
  \item{\code{statsQC}}{Getter for the quality control statistics}
}
}

\examples{

}
\author{
Avishay Spitzer
}
