% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{scandal_plot}
\alias{scandal_plot}
\title{}
\usage{
scandal_plot(p, show_plot = TRUE, save_to_file = FALSE,
  project_dir = NULL, plots_dir = "plots", filename = NULL,
  dirname = NULL, device = "png")
}
\arguments{
\item{device}{}
}
\value{

}
\description{

}
\details{

}
\author{
Avishay Spitzer
}
