% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{logtpm}
\alias{logtpm}
\alias{logtpm<-}
\alias{qualityControl}
\alias{qualityControl<-}
\alias{nodeID}
\alias{nodeID<-}
\alias{projectID}
\alias{sampleIDs}
\alias{unprocessedData}
\alias{preprocConfig}
\alias{inspectSample}
\alias{cell2SampleMap}
\title{ScandalDataSet generics}
\usage{
logtpm(object, ...)

logtpm(object, ...) <- value

qualityControl(object, ...)

qualityControl(object, ...) <- value

nodeID(object, ...)

nodeID(object, ...) <- value

projectID(object, ...)

sampleIDs(object, ..., return_sorted = FALSE)

unprocessedData(object, ...)

preprocConfig(object, ...)

inspectSample(object, nodeID, ...)

cell2SampleMap(object, ...)
}
\description{
ScandalDataSet generics
}
