% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{logtpm}
\alias{logtpm}
\alias{logtpm<-}
\alias{qualityControl}
\alias{qualityControl<-}
\alias{nodeID}
\alias{projectID}
\alias{nodeIDs}
\alias{unprocessedData}
\alias{preprocConfig}
\alias{inspectNode}
\title{ScandalDataSet generics}
\usage{
logtpm(object, ...)

logtpm(object, ...) <- value

qualityControl(object, ...)

qualityControl(object, ...) <- value

nodeID(object, ...)

projectID(object, ...)

nodeIDs(object, ..., return_sorted = FALSE)

unprocessedData(object, ...)

preprocConfig(object, ...)

inspectNode(object, nodeID, ...)
}
\description{
ScandalDataSet generics
}
