% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{ScandalDataSet}
\alias{ScandalDataSet}
\title{}
\usage{
ScandalDataSet(..., childNodes = S4Vectors::SimpleList(),
  parentNode = NULL, identifier = exp_id(),
  preprocConfig = DefaultPreprocConfig())
}
\arguments{
\item{preprocConfig}{}
}
\value{
A ScandalDataSet object is returned from the constructor
}
\description{

}
\details{

}
\author{
Avishay Spitzer
}
