% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ScandalDataSet-class}
\alias{ScandalDataSet-class}
\alias{ScandalDataSet}
\alias{logtpm,ScandalDataSet-method}
\alias{logtpm<-,ScandalDataSet-method}
\alias{childNodes,ScandalDataSet-method}
\alias{childNodes<-,ScandalDataSet-method}
\alias{parentNode,ScandalDataSet-method}
\alias{parentNode<-,ScandalDataSet-method}
\alias{nodeID,ScandalDataSet-method}
\alias{projectID,ScandalDataSet-method}
\alias{sampleNames,ScandalDataSet-method}
\alias{unprocessedData,ScandalDataSet-method}
\alias{preprocConfig,ScandalDataSet-method}
\title{ScandalDataSet class}
\usage{
## Constructor
ScandalDataSet(..., childNodes = S4Vectors::SimpleList(),
  parentNode = NULL, preprocConfig = DefaultPreprocConfig(),
  nodeID = NODE_ID(), projectID = PROJ_ID())

\S4method{logtpm}{ScandalDataSet}(object, ...)

\S4method{logtpm}{ScandalDataSet}(object, ...) <- value

\S4method{childNodes}{ScandalDataSet}(object)

\S4method{childNodes}{ScandalDataSet}(object) <- value

\S4method{parentNode}{ScandalDataSet}(object)

\S4method{parentNode}{ScandalDataSet}(object) <- value

\S4method{nodeID}{ScandalDataSet}(object)

\S4method{projectID}{ScandalDataSet}(object)

\S4method{sampleNames}{ScandalDataSet}(object)

\S4method{unprocessedData}{ScandalDataSet}(object)

\S4method{preprocConfig}{ScandalDataSet}(object)
}
\arguments{
\item{...}{arguments to pass to the \code{SingleCellExperiment} constructor.}

\item{childNodes}{a \code{SimpleList} object containing the child nodes of the
constructed \code{ScandalDataSet} object. Default is an empty \code{SimpleList}}

\item{parentNode}{the parent node of the constructed \code{ScandalDataSet}
object. Default is \code{NULL}, meaning that the constructed object has no
parent.}

\item{preprocConfig}{a configuration object of class \code{PreprocConfig}}

\item{nodeID}{a unique identifier of the constructed \code{ScandalDataSet}
object that should represent the specific sample. If not supplied a unique ID
will be generated randomly however it is advised to set this field.}

\item{projectID}{an identifier common to all the nodes in the constructed
\code{ScandalDataSet} object.  If not supplied a unique ID
will be generated randomly however it is advised to set this field.}

\item{value}{a value to replace the currently set value (applies to all setter methods).}
}
\description{
An S4 class for storing single-cell seqeuncing data, analysis
}
\details{
The S4 class \code{ScandalDataSet}
}
\section{Slots}{

\describe{
\item{\code{childNodes}}{a \code{SimpleList} object containing the child nodes of the
constructed \code{ScandalDataSet} object.}

\item{\code{parentNode}}{the parent node of the constructed \code{ScandalDataSet}
object.}

\item{\code{unprocessedData}}{a read-only matrix that contains the unprocessed data that
allows re-accessing this data without the need to read it from file. Sparse matrix
representation as well as maintaining a single copy for the entire objects tree
decreases the memory overhead of this approach.}

\item{\code{preprocConfig}}{a configuration object of class \code{PreprocConfig}.}

\item{\code{nodeID}}{a unique character identifier of the constructed \code{ScandalDataSet}
object that should represent the specific sample.}

\item{\code{projectID}}{a character identifier common to all the nodes in the constructed
\code{ScandalDataSet} object.}
}}

\section{Constructor}{

Constructs a \code{ScandalDataSet} object.
}

\section{Methods}{

\describe{
  \item{\code{logtpm}}{Getter/setter for the logtpm assay}
  \item{\code{childNodes}}{Getter/setter for childNodes}
  \item{\code{parentNode}}{Getter/setter for the parentNode}
  \item{\code{unprocessedData}}{Getter for the unprocessedData (read-only)}
  \item{\code{preprocConfig}}{Getter for the preprocConfig (read-only)}
  \item{\code{nodeID}}{Getter/setter for the nodeID}
  \item{\code{projectID}}{Getter/setter for the projectID}
}
}

\examples{

}
\author{
Avishay Spitzer
}
